Refactoring ConnectionModal
